Hack-and-slash port to nasm and Windows
Semi-related, but not particularly. Instead of targeting gas asm
it now generates nasm syntax (which is Intel style, not AT&T).

All tests in the test suite pass, including driver.sh, however
test-stage2 still needs some work.

Additionally, it can now be built and run on Windows, with many
limitations. All of the driver functionality is missing when running on
Windows; it only supports a single .c to a single .s file. And, while
the compiler runs on Windows, it's a to-Linux cross-compiler and
generates a .s meant to be linked there, uses SysV ABI, etc.