A few more fixes to get test-stage2 working with nasm
It seems that nasm confuses references to an extern named "wait"
with the wait instruction (at least in some contexts) so add a
workaround for that for now (compiled by host compiler). Posted
to nasm-devel, but I'm not sure about a patch to nasm yet.

Additionally fix an error in common symbols (two statics named
output_file, oops!) and few more att<->nasm syntax changes and
now test-stage2 is all working.