Port to DynASM, passes (almost) all selftests
Changes main output to be generating machine code via DynASM, rather
than outputting text to be passed to the assembler.

Because chibicc assumes run-and-exit, it doesn't do any free/cleanup.
To work around this, while compiling multiple translation units in a
single run, each module jankily purges and resets and the heap is tossed
between files.

For each translation unit, a custom-obj-ish thing is output (which maybe
should have been ELF or whatever instead of custom...). After compiling
all the files listed on the command line, it does a link pass to handle
function references cross-module as well as data relocations (for
example to create literals or initialize global data). There's also no
particularly good reason that these objs (.dyo) need to be written to
disk (rather than simply being an in-memory structure), it was just
useful to debug the output.

All chibicc self-tests are passing except for two: 1) asm.c (because it
just passes the text of the asm block to the system assembler which no
longer exists); and 2) tls.c just because TLS support hasn't been
written yet.

This change also deletes most of the "driver" part of main.c as it
didn't seem to make a lot of sense with a JIT mode. Perhaps it would be
useful to bring some of that back, if only to be able to use it to build
common C software like sqlite, etc. for testing purposes.

The code for -S (output textual assembler) that is chibicc's normal mode
is still in the code, this was done during the rewrite to make it easier
to track. A future change will remove that code; a pristine copy of
chibicc is equally useful to compare final output with in any case.