Add ASAN hooks to poison bumpalloc heap, fix RDI use on Win
When running under -fsanitize=address, hook into the poison/unpoison
runtime routines to get bounds checking (and leak checking on Linux).

Using ASAN on Windows caused a mostly-unrelated crash caused by
incorrect use of RDI during codegen -- I was treating it as volatile,
and a utility register, but on Windows it's non-volatile which the ASAN
runtime cleanup after main() happened to rely on due to lucky
optimizer-chance. (It's kind of unfortunate for 'rep stosb' usage too,
which is used to initialize locals, perhaps should revisit that.) RCX is
now used on Windows instead for a default utility volatile register.