Support a basic form of relinking
By passing in a previous LinkInfo to compile_and_link, data is not
recreated or reinitialized, but code is relinked (and fixed up to point
at the previously allocated data). This allows for a simple version of
hot reloading for experimentation. Only manual testing so far (in
another code base using dyibicc); a higher-level test driver that
sequences and tests the results of code changes is needed next.