Tag string constants as 'rodata'
This is used as a signal during relinking that the can be updated to
what's in the source code rather than maintained as user data that
should be saved post relinking.

Add a copy of khash.h too, as hashmap.c only does string keys and this
needed some void* keys. (It's a bit awkward to use, possibly look at
building and wrapping swisstables or something.)