Move all compiler and linker into two state objects
Removes all per-file state (including heap allocation) and merges into a
single location for clearer lifetime.

This introduces two heaps and lifetimes for allocations, scoped to the
compiler lifetime, the linker (plus the lifetime of all the compiler
invocations). Additionally, a third type "manual" is for allocations
that include allocation of code/data segments by the linker for
returning to the caller, as well as metadata about them for subsequent
relinks.

In future changes, the goal will be to avoid all abort() and exit()
calls by longjmp'ing back to the embedding entry point and being able to
reset compiler/linker state by memset'ing these two state structures.