Recover some of the performance lost after moving into state obj
https://github.com/sgraham/dyibicc/commit/9d8aa1155560bb247bcf3643f24884238b5b42c2
regressed performance noticeably. The biggest problem was way too much
accidental memset'ing of the heap, but the meat of the change that moved
state from file local statics into a single global object also seems to
hurt.

I briefly investigated whether simply /GL+/LTCG would recover the
difference, but it does not.

Before combining:
/Ox: 2.8s
/Ox /GL: 2.35s

After combining:
/Ox: 3.2s
/Ox /GL: 2.75s

That is, /GL can make up for what was lost, but it also makes the
non-combined version faster too.

This change fixes the memset bug, but the next step is to revert this
one and its predecessor (the state combining change), and work on
getting a better (longer running) benchmark. Then we can go back to
figuring out the embedding/error reporting situation with that
available.