Let data initializations see all functions as well
And also add many more to the hacky link.c list (that really needs to be
GetProcAddress'ing on some dlls instead).

With this, the sqlite3 amalgamation of shell.c and sqlite3.c builds and
starts running, but hits one of the abort()ing unimplemented intrinsic
calls before anything interesting happens.