Fix compile-time regression in alloc
The regression had nothing to do with the wrapping of state in a global
struct... which should have been obvious, but it took some staring at
disassembly to convince myself.

My other excuse is that when you run it in a profiler, the regression
doesn't show up... because it happens before main and in the loader.

Because I embedded the heaps into the state object, they (of course, in
retrospect) become .bss so they have to be cleared before the program
starts, which ain't free. Such a facepalm. Instead go back to
VirtualAlloc'ing the heaps, and all is well.

On the plus side, I am now confident that there's no cost to the global
state object (at least in /Ox /GL), so embedding improvements can merge
and continue.