Merge inmemlink-2, removes use of .dyo files
The compile process (can be) now fully fileless. Typical command line
operation reads from .c/.h, but the code is generated (and linked)
directly in memory, rather than going through the intermediate step of
dyo files being written and re-read.

This is slightly simpler in that it avoids the de/serialization code,
but the memory management is a bit more complex.

A subsequent change will probably need to reintroduce some sort of
textual dump to inspect compile/fixups/link for debugging, but it wasn't
necessary yet.

Additionally, symbols are maintained as regular strings in the fixup
array and some hashtables. It would be nice to intern these to speed up
and simplify linking.