Build with ninja instead of make/nmake
I was getting fed of up tweaking CFLAGS and `make clean`s. Tests are
also run by ninja instead of the minilua script, so revert additions to
that. The tests are normally run in parallel now by ninja which is
possible now that there's no clashing .dyo files being generated. It's
also much faster if you have lots of cores.

There are 3 configs, 'd' for Debug, 'r' for Release, and 'a' for ASAN,
and they're in separate out dirs, e.g. `out/lr` for "L"inux "R"elease.