Add script to amalgamate the compiler into a single .c file
src/build_amalg.py generates embed/libdyibicc.{c,h} which are a
standalone copy of the compiler for simple embedding.

There's some clutter added to the normal source build (which is defined
away) in order to make all functions (and data) except the four exported
in libdyibicc.h be static.