Add helpers and first automated test of update functionality
The main compiler binary and tests only ran in a "normal" compiler batch
mode style.

This changes adds a python scripting helper that generates .c runners
that build against the amalgamated build and simulate running a number
of sequential updates to the code by calling dyibicc_update() and
calling the entry point repeatedly.

This should help cover regressions in update() but also
configuration/warning issues in the amalgamated build.

It's a little bit circular/messy right now because building the
amalgamated build isn't part of the main ninja run. In order to get the
update tests to run against HEAD of src/, src/build_amalg.py needs to be
run manually first (which in turn also requires running `m r` and `m.bat
r`). This should be improved.