Add an update test for multiple files
This required a bit of adjustment to the file loading and how the
initial update() is called. I don't really like how there's an asymmetry
in how file contents are loaded at different times, but I also don't
want update to have to do a callback for the contents for no reason, so
this is how it is for now.

libdyibicc.h got more documentation to explain that problem, and is now
also generally more helpful.

Additionally, actually testing two file updates exposed that it didn't
work! The not-being-updated files weren't being
VirtualProtect()/mprotect()'d back to writeable during relink, so it'd
crash because the memory was only R+X after the first link.