Move build of amalgamated into main ninja as it's used for tests now
Still kind of hacky, but at least it's used in the normal run of `m d
test` so problems should be more visible.

The update tests also uncovered leaks (asan under linux) in global data
and exports, but I think they're going to be a bit involved to fix
properly so deferring to next change.