Two unrelated Win ABI fixes for struct passing
The first was happening calling raylib math functions which tend to pass
vectors and matrices by value. The extra stack space for making a copy
to pass-by-value as far as C is concerned (which is then passed by
reference at the ABI level!) was not being accounted for in keeping the
stack 16 byte aligned. This would sometimes manifest in a crash (which
seemed to be more common in /Ox builds, but not for any good reason.

While testing this, struct_test6 which passes Ty6, which is a struct
containing a single array of char[3] started failing for no obviously
related reason, and even now I think it was just chance. A struct of
char[3] is not passed by register, and this bug was on the "receiving"
side, not the calling side -- it was assuming that the 3 bytes of the
struct would be in a register, but they won't be, the first argument
will be a pointer to the address where the 3 bytes are found.