Emit RUNTIME_FUNCTION .pdata on Windows
Normally stored in the .pdata section in the PE. These allow unwinding
the stack properly during exceptions, which we'll be using in the future
to guard against general GPFs. They're also used for general stack
unwinding which makes windbg/VS display a correct call stack when
stepping through disassembly

This is still a partial implementation:
- Functions that allocate > 4k and need a _chkstk don't have an entry
  emitted yet.
- In an exception, I think rdi (non-volatile on Windows) wouldn't be
  restored properly. It's locally saved/restored rather than in the
  function prolog, and should probably move. On the other hand, it's
  only used for clearing locals, so hopefully not a likely source of
  exceptions.

There's also no tests because it's mostly about the UI working properly
in VS right now.