Merge pdb branch, adds pdb writing library
This was substantially harder than I expected when I tried. The library
generates a .pdb containing symbol and line debugging information, as
well as some hackery to generate a stub .dll that convinces Visual
Studio to load the generated .pdb.

The pdb writer is dyn_basic_pdb.h and is standalone. There's an example
of using it in dbp_example/ with a Visual Studio solution (but it should
build easily directly.) dyn_basic_pdb.h isn't used by dyibicc yet, but
that's coming next.