Set up fuzzing
Add a separate entry point for libFuzzer, and then a bit assistance in
to build a fuzz config, and then some test infrastructure to test the
inevitable first simple errors it found.

Special test running is required because the first one it found was
related to not having a newline at the end of an erroneous preprocessor
line, so the files need to be passed verbatim from the fuzzer, rather
than with additional "how to run", and "what to expect" comments added.

I think libFuzzer only works on Linux, so `./m f` only works there. I
tried afl-fuzz too, but afl-clang seems to mis-compile (?) the compiler,
or otherwise mess up something with mmap(), so for now just using
libFuzzer's.