Integrate pdb generation
Previously dyn_basic_pdb.h was added but only used by its example
program. Now, when running on Windows and with -g passed on the command
line, dyibicc will generate pdb and cause an attached debugger to load
the symbols. The .pdata/.xdata doesn't seem to be working correctly,
despite all the mucking around I did for that in dbp, that will need to
be debugged next.

There's also a bit more #if X64WIN than I'd like, I guess I should emit
similar data for Linux.

This also makes plain how awful the "stack machine" style codegen is
(when looking at mixed disasm view). It really hasn't been a problem
speed-wise for playing in Rdy, but it would be nice to remove some of
the offensively redundant instructions, or try to rework for
"Destination-Driven Code Generation".