Fix referencing passed-by-ref structs in win abi
The double-dereference necessary to go from the function argument to the
actual location (in the case of a large structure pass) was incorrectly
happening in the field codegen. This worked if the structure was only
ever referenced as `obj.xyz`, but if `obj` is directly used (for
example, but assignment) then the address wasn't correctly calculated.